Create toolbar

+++ app/src/main/java/com/yaroshevich/podacha/room/dao/PanelDao.kt
@@ -3,7 +3,6 @@ package com.yaroshevich.podacha.room.dao
 import androidx.lifecycle.LiveData
 import androidx.room.*
 import com.yaroshevich.podacha.room.entities.Panel
-import com.yaroshevich.podacha.room.entities.Session
 
 @Dao
 interface PanelDao {
@@ -12,7 +11,7 @@ interface PanelDao {
     fun getAll(): LiveData<List<Panel>>
 
     @Query("SELECT * FROM panel WHERE id = :id")
-    fun getById(id: Long): Panel
+    fun getById(id: Int): Panel
 
     @Insert
     fun insert(panel: Panel)
